% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{aggCells}
\alias{aggCells}
\title{Aggregate cellular HDD/CDD values to country-wide average (population-weighted)}
\usage{
aggCells(data, weight, mask)
}
\arguments{
\item{data}{raster object containing HDD/CDD values}

\item{weight}{raster object containing aggregation weights}

\item{mask}{raster opbject defining (regional) aggregation boundaries}
}
\value{
data frame containing regionally meaned HDD/CDD values
}
\description{
Aggregate cellular HDD/CDD values to country-wide average (population-weighted)
}
