% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggToLevel.R
\name{aggToLevel}
\alias{aggToLevel}
\title{Aggregate variables with respect to a specified level}
\usage{
aggToLevel(df, variables, level, newname)
}
\arguments{
\item{df}{A data frame containing the data to be aggregated.}

\item{variables}{A character vector of variable names that should be aggregated.}

\item{level}{A string specifying the column in \code{df} that defines the level at which
aggregation is performed.}

\item{newname}{A string specifying the name for the newly aggregated category.}
}
\value{
A data frame where the selected variables have been aggregated into
a single variable with the specified \code{newname}.
}
\description{
This function aggregates selected variables at a given categorical level and
replaces them with a new summarized entry in the dataset.
}
