% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{blend}
\alias{blend}
\title{Weighted blend of BAIT and near-surface atmospherical temperature}
\usage{
blend(bait, tas, weight)
}
\arguments{
\item{bait}{raster data on BAIT}

\item{tas}{raster data on near-surface atmospherical temperature}

\item{weight}{named list with blending parameters bLower, bUpper, bMax}
}
\value{
blended raster data
}
\description{
To adress loss of buildings' internal memory of previous conditions at high
outside temperatures due to window opening, etc., BAIT and outside temperature
are blended. The blend is active between a lower and upper temperature threshold,
\code{bLower} and \code{bUpper}, which are mapped to a range of -5 to +5 of a
sigmoid function (corresponding to a 1\% and 99\% blend). The maximum amount of
blending (i.e. the amplitude of the sigmoid function) is given by a parameter \code{bMax}.
}
