% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcACOwnershipRates.R
\name{calcACOwnershipRates}
\alias{calcACOwnershipRates}
\title{Calculate Air Conditioning Ownership Rates}
\usage{
calcACOwnershipRates()
}
\value{
A data frame with one data point per region
}
\description{
Calculates air conditioning ownership rates per household by combining
data from household surveys, IEA, and Odyssee. Data sources are prioritized in the
following order: Odyssee > IEA > household surveys.
}
\author{
Hagen Tockhorn
}
