% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBAITpars.R
\name{calcBAITpars}
\alias{calcBAITpars}
\title{calculate regression parameters for BAIT climate variables}
\usage{
calcBAITpars(model = "GFDL-ESM4")
}
\arguments{
\item{model}{specify GCM responsible for data input}
}
\value{
terra SpatRaster covering one regression parameter per layer per cell
}
\description{
linear regression on historic data to determine regression
parameters for surface downdwelling shortwave radiation (rsds), near-surface
wind speed (sfcwind) and near-surface specific humidity (huss) with respect
to near-surface air temperature (tas).
The regression is done with on a simple linear model, where the historical
input data covers the years of 2000-2014. For rsds and sfcwind, a simple linear
relationship is assumed where for huss an exponential relation is assumed,
buildig upon the non-linear relation between water vapor pressure and temperature.
}
\author{
Hagen Tockhorn
}
