% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBRG.R
\name{calcBRG}
\alias{calcBRG}
\title{Calculate total boiler sales from BRG}
\usage{
calcBRG()
}
\value{
MagPIE object with total boiler sales in European countries
}
\description{
BRG has extremely comprehensive (and extremely expensive) sales data. But one
can recover total sales of boilers for space heating from the EHPA report.
}
\author{
Robin Hasse
}
