% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBuildingStock.R
\name{calcBuildingStock}
\alias{calcBuildingStock}
\title{calculate the European historic building stock}
\usage{
calcBuildingStock(subtype = c("residential", "commercial"))
}
\arguments{
\item{subtype}{Character with subsector}
}
\value{
MAgPIE object with historic building stock
}
\description{
Join data from various sources to construct a consistent European building
stock with a focus on floor space (million m2).
}
\note{
if the distribution of building types is missing, it is filled with the
Europe-wide distribution. It might be worthwhile to map filling countries
instead.
}
\author{
Robin Hasse
}
