% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarrierPrices.R
\name{calcCarrierPrices}
\alias{calcCarrierPrices}
\title{Calculate carrier prices}
\usage{
calcCarrierPrices()
}
\description{
Final energy carrier prices including energy supply cost, transport and
distribution cost(T&D) and taxes (without VAT). CO2 price component and VAT
are added later depending on the scenario assumptions.
}
\author{
Robin Hasse
}
