% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{calcCellHDDCDD}
\alias{calcCellHDDCDD}
\title{Assign HDD/CDD values for given ambient/limit temperature}
\usage{
calcCellHDDCDD(temp, typeDD, tlim, factors)
}
\arguments{
\item{temp}{raster data containing temperature/BAIT values}

\item{typeDD}{type of degree day}

\item{tlim}{limit temperature}

\item{factors}{data frame with degree day values for \code{temp/tlim} combination}
}
\value{
raster object with HDD/CDD values
}
\description{
Assign HDD/CDD values for given ambient/limit temperature
}
