% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostConstruction.R
\name{calcCostConstruction}
\alias{calcCostConstruction}
\title{calculate construction cost}
\usage{
calcCostConstruction(granularity = NULL)
}
\arguments{
\item{granularity}{character, name of BRICK granularity}
}
\description{
floor-space specific construction cost
}
\author{
Robin Hasse
}
