% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostRenovation.R
\name{calcCostRenovation}
\alias{calcCostRenovation}
\title{calculate renovation cost}
\usage{
calcCostRenovation(energyLadder = FALSE)
}
\arguments{
\item{energyLadder}{logical, should the calculation include renovation
transitions that are considered a decline on the energy ladder?}
}
\value{
MagPIE object with floor-space specific renovation cost depending on
the initial and final state of the building
}
\description{
floor-space specific renovation cost
}
\author{
Robin Hasse
}
