% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostRenovationBS.R
\name{calcCostRenovationBS}
\alias{calcCostRenovationBS}
\title{calculate renovation cost}
\usage{
calcCostRenovationBS(granularity = NULL)
}
\arguments{
\item{granularity}{character, name of BRICK granularity}
}
\value{
MagPIE object with floor-space specific renovation cost depending on
the initial and final state of the building
}
\description{
floor-space specific renovation cost
}
\author{
Robin Hasse
}
