% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostRenovationHS.R
\name{calcCostRenovationHS}
\alias{calcCostRenovationHS}
\title{calculate renovation cost}
\usage{
calcCostRenovationHS(energyLadder = FALSE, granularity = NULL)
}
\arguments{
\item{energyLadder}{logical, should the calculation include renovation
transitions that are considered a decline on the energy ladder?}

\item{granularity}{character, name of BRICK granularity}
}
\value{
MagPIE object with floor-space specific renovation cost depending on
the initial and final state of the building
}
\description{
floor-space specific renovation cost
}
\author{
Robin Hasse
}
