% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDensity.R
\name{calcDensity}
\alias{calcDensity}
\title{Get Population Density}
\usage{
calcDensity(endOfHistory = 2025)
}
\arguments{
\item{endOfHistory}{upper temporal boundary for historical data}
}
\value{
magpie object
}
\description{
Get Population Density
}
\author{
Hagen Tockhorn
}
