% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiscountRate.R
\name{calcDiscountRate}
\alias{calcDiscountRate}
\title{Discount rate}
\source{
https://doi.org/10.1016/j.eneco.2011.07.010
}
\usage{
calcDiscountRate(granularity = NULL)
}
\arguments{
\item{granularity}{character, name of BRICK granularity}
}
\description{
Discount rate including implicit discounting
}
\author{
Robin Hasse
}
