% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEfficiencyRegression.R
\name{calcEfficiencyRegression}
\alias{calcEfficiencyRegression}
\title{Calculation Regression Parameters for FE-UE Efficiency Projection}
\usage{
calcEfficiencyRegression(GasBioEquality = TRUE)
}
\arguments{
\item{GasBioEquality}{Determines if Natgas and Biomod share the same efficiencies}
}
\description{
Calculate the regression parameters for the projection of final (FE)
to useful (UE) energy conversion projection for all combinations of
energy enduses and carriers. The regression parameters correspond to an
asymptotic regression model and encompass the parameters Asym, R0 and lrc.
The are determined using a nonlinear least-squares regression.
}
\details{
This approach closely follows the model by De Stercke et al. which is
mainly driven by GDP per Capita.
}
\references{
De Stercke, S. (2014). Dynamics of Energy Systems: A Useful
Perspective (Interim Report, p. 68). IIASA.
http://pure.iiasa.ac.at/id/eprint/11254/1/IR-14-013.pdf
}
\author{
Hagen Tockhorn
}
