% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEdemandBuildings.R
\name{calcFEdemandBuildings}
\alias{calcFEdemandBuildings}
\title{calculate FE demand in buildings}
\usage{
calcFEdemandBuildings()
}
\description{
Simple estimate of total FE demand in differet building types
}
\details{
Schould only be used as weights not actual demand. The disaggregation to
building types is too simplistic.
}
\author{
Robin Hasse
}
