% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEUE.R
\name{calcFEUE}
\alias{calcFEUE}
\title{Historical Final and Useful Energy Demand}
\usage{
calcFEUE()
}
\value{
data.frame with final and useful energy demand
}
\description{
Historic Final Energy (FE) Demand and FE-EU-Efficiencies are taken to
compute Useful Energy (UE) Demand.
}
\author{
Hagen Tockhorn
}
