% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEUEefficiencies.R
\name{calcFEUEefficiencies}
\alias{calcFEUEefficiencies}
\title{Calculation and Projection of Final to Useful Energy}
\usage{
calcFEUEefficiencies(gasBioEquality = TRUE)
}
\arguments{
\item{gasBioEquality}{Determines if carriers natgas and biomod share the same efficiencies}
}
\description{
Calculate Efficiencies of Final (FE) to Useful (UE) Energy Conversion for all
combinations of Energy Carriers and Enduses.
The efficiency projections are based on a model by De Stercke et al. which is
mainly driven by GDP per Capita. It describes an S-shaped curve approaching
assumed efficiency levels. The parameters of that curve are derived by a
regression with observations of IEA data.
}
\references{
De Stercke, S. (2014). Dynamics of Energy Systems: A Useful
Perspective (Interim Report, p. 68). IIASA.
http://pure.iiasa.ac.at/id/eprint/11254/1/IR-14-013.pdf
}
\author{
Hagen Tockhorn
}
