% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspacePerCap.R
\name{calcFloorspacePerCap}
\alias{calcFloorspacePerCap}
\title{Residential floor space per capita}
\usage{
calcFloorspacePerCap(scenario)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned.}
}
\value{
MagPIE object with residential floor space per capita
}
\description{
Residential floor space per capita
}
\author{
Robin Hasse
}
