% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{calcHDDCDD}
\alias{calcHDDCDD}
\title{Calculate Heating and Cooling Degree Days (HDD/CDD)}
\usage{
calcHDDCDD(scenario = c("ssp2"), fromSource = TRUE)
}
\arguments{
\item{scenario}{Character vector specifying the socio-economic scenario(s), e.g., \code{"ssp2"}.}

\item{fromSource}{Logical. If \code{TRUE}, loads HDD/CDD data from source.
If \code{FALSE}, computes them using the CLIMBED package.}
}
\value{
A magclass object with annual degree-day values.
}
\description{
Computes or loads annual heating and cooling degree days (HDDs and CDDs)
for given socio-economic scenarios. If \code{readFromSource = TRUE},
data is read from the \code{HDDCDD} source. Otherwise, degree days are
calculated using the \strong{CLIMBED} package with specified parameters
such as temperature limits and standard deviations.
}
\author{
Hagen Tockhorn
}
