% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeatingCapacity.R
\name{calcHeatingCapacity}
\alias{calcHeatingCapacity}
\title{Floor-area specific heating capacity}
\source{
https://www.heizung.de/ratgeber/diverses/heizleistung-berechnen-gruende-und-ablauf.html
}
\usage{
calcHeatingCapacity(swissFormular = FALSE)
}
\arguments{
\item{swissFormular}{boolean, apply the simplistic swiss formular?}
}
\value{
MagPIE object with floor-area specific heating capacity
}
\description{
Rough assumption based on online reference values scaled with U-values from
Hotmaps and heating degree days (HDD).
}
\details{
This is a very rough estimation that neglects regional differences and many
other effects. It should be replace by bottom-up calculation based on climate
data.
}
\author{
Robin Hasse
}
