% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeatingSystemReplacement.R
\name{calcHeatingSystemReplacement}
\alias{calcHeatingSystemReplacement}
\title{Calculate Matrix of heating system replacement shares}
\usage{
calcHeatingSystemReplacement()
}
\value{
MagPIE object with shares
}
\description{
Calculate Matrix of heating system replacement shares
}
\author{
Robin Hasse
}
