% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeatingSystemSales.R
\name{calcHeatingSystemSales}
\alias{calcHeatingSystemSales}
\title{Calculate yearly sales of heating systems}
\usage{
calcHeatingSystemSales()
}
\value{
MagPIE object with heating system sales
}
\description{
We use national data from industry association for the biggest European
countries and European data sets from the EHI and EHPA.
}
\author{
Robin Hasse
}
