% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHouseholdSize.R
\name{calcHouseholdSize}
\alias{calcHouseholdSize}
\title{calculate household size}
\usage{
calcHouseholdSize()
}
\value{
MAgPIE object with household size
}
\description{
Number of persons per household.
}
\note{
THE EU Buildings Database gives one share of all dwellings with 5 or
more persons. This share is weighted with 5 persons which can cause a
downward bias.
}
\author{
Robin Hasse
}
