% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEAPFU.R
\name{calcIEAPFU}
\alias{calcIEAPFU}
\title{IEA database of primary, final and useful energy}
\usage{
calcIEAPFU()
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes data for the Primary Final Useful (PFU) resolution
}
\examples{
\dontrun{ a <- calcOutput("IEA_PFU")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Antoine Levesque, Robin Krekeler
}
