% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLifetimeParams.R
\name{calcLifetimeParams}
\alias{calcLifetimeParams}
\title{Weibull lifetime distribution parameters}
\source{
https://www.eia.gov/analysis/studies/buildings/equipcosts/pdf/full.pdf

http://dx.doi.org/10.1016/j.enbuild.2017.01.080
}
\usage{
calcLifetimeParams(subtype)
}
\arguments{
\item{subtype}{character, type of asset (either \code{building}, 'heatingSystem'
or 'buildingShell')}
}
\value{
MagPIE object with Weibull lifetime distribution parameters
}
\description{
Parameters for the lifetime of heating systems are taken from a very
detailed EIA publication for building sector appliances and equipment. The
range of the building shell lifetime is taken from Skarning et al. 2017.
}
\author{
Robin Hasse
}
