% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLTRS.R
\name{calcLTRS}
\alias{calcLTRS}
\title{Calculate data based on Long Term Renovation Strategies}
\usage{
calcLTRS(refGranularity = c("typ", "sec"))
}
\arguments{
\item{refGranularity}{character, for \code{"typ"}, you get the most granular
data with residential buildings split by SFH and MFH. However, this data is
less reliable and incomplete. \code{"sec"} will aggregate residential
buildings to Res. It should be complete for EU27.}
}
\value{
MagPIE object
}
\description{
Extrapolate the residential heating system distribution in each vintage from
DEU to other countries and to commercial buildings.
}
\details{
We use iterative proportional fitting to extrapolate the German residential
pattern to other countries and subsectors. This data is extremely rough and
should be taken with care. It is helpful in the matching nonetheless to
represent interactions between vintage and heating systems while other more
reliable references assure overall consistency.
}
\author{
Robin Hasse
}
