% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMatchingReference.R
\name{calcMatchingReference}
\alias{calcMatchingReference}
\title{Calculate Reference data for input matching}
\usage{
calcMatchingReference(subtype)
}
\arguments{
\item{subtype}{character, matching reference}
}
\description{
Calculate Reference data for input matching
}
\author{
Robin Hasse
}
