% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOdysseeStock.R
\name{calcOdysseeStock}
\alias{calcOdysseeStock}
\title{Calculate Odyssee building stock variables}
\usage{
calcOdysseeStock(interpolate = FALSE)
}
\arguments{
\item{interpolate}{logical, if TRUE, floor space growth rates from JRC_IDEES
are used to inter- and extrapolate the data}
}
\value{
MagPIE object with building stock data
}
\description{
Residential stock variables are given in number of dwellings, flats or
houses while commercial stock is given in Floor space. This function
computes all stock variables in terms of floor space considering dwelling
size.
}
\details{
The data can be inter- and extrapolated using IDEES floor space data. As we
trust the absolute level less than Odyssee, we only consider the growth to
fill values.
}
\author{
Robin Hasse
}
