% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPFUDB.R
\name{calcPFUDB}
\alias{calcPFUDB}
\title{Converts Data from Primary Fuel and Useful Energy Data Base (PFUDB)}
\usage{
calcPFUDB()
}
\value{
magpie object
}
\description{
PFU Data is disaggregated into different thermal and non-thermal Enduses
using FE Shares from Odyssee and IEA_ETP.
}
\details{
This was adapted from EDGE function 'getPFUDB.R'.
}
\author{
Hagen Tockhorn
}
