% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationBuildings.R
\name{calcPopulationBuildings}
\alias{calcPopulationBuildings}
\title{Historic and projected population}
\usage{
calcPopulationBuildings(granularity = NULL)
}
\arguments{
\item{granularity}{character, name of BRICK granularity}
}
\value{
MagPIE object with population at BRICK granularity
}
\description{
Population differentiated by location (urban/rural) and building type
}
\author{
Robin Hasse
}
