% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRenDepth.R
\name{calcRenDepth}
\alias{calcRenDepth}
\title{Calculate renovation depth}
\usage{
calcRenDepth()
}
\description{
The renovation depth is currently calculated based on the ratio of specific
useful energy demand after and before the renovation. This is valid under the
assumption that there is no change in behaviour (rebound effect) which we
currently make. Once adaptive behaviour is considered, we should use the
ratio of U-values.
}
\author{
Robin Hasse
}
