% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRenovation.R
\name{calcRenovation}
\alias{calcRenovation}
\title{Calculate historic renovation flows}
\usage{
calcRenovation()
}
\value{
MagPIE object with historic renovation flows
}
\description{
Combine data on heating system sales and replacement to derive full
building state transition data
}
\author{
Robin Hasse
}
