% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRenovationCostModel.R
\name{calcRenovationCostModel}
\alias{calcRenovationCostModel}
\title{Calculate linear renovation cost model}
\source{
https://op.europa.eu/s/xnYt
}
\usage{
calcRenovationCostModel()
}
\value{
MAgPIE object with specific renovation cost
}
\description{
Specific renovation cost are estimated with a linear model with intercept
w.r.t. to the renovation depth.
}
\details{
The specific investment cost for renovation depending on the depth of
renovation is calculated using a two-level statistical model fitted with
data from a European Commission (EC) report on renovation.
In a fist step, we fit a simple linear model with intercept to predict
specific investment (USD/m2) with relative PE savings. In a second step, this
linear model is scaled with one factor for each region. This scaling factor
is predicted with GDP/POP using a negative exponential curve. For EU regions
that are reported, we correct the result with a region-specific but
time-invariant factor to best match the linear model to the data. This way,
we get a linear model with intercept for each region that can be extrapolated
to other regions (using GDP/POP) which still matches data of EU regions very
well. The cost is finally disaggregated across residential building types
(SFH, MFH) based on a rough cost factor seen in the ENTRANZE data.
}
\author{
Robin Hasse
}
