% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareDemolition.R
\name{calcShareDemolition}
\alias{calcShareDemolition}
\title{Calculate minimum share of demolished buildings}
\usage{
calcShareDemolition()
}
\description{
Calculate minimum share of demolished buildings
}
