% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareETP.R
\name{calcShareETP}
\alias{calcShareETP}
\title{Share of energy carriers or end uses in historic buildings demand}
\usage{
calcShareETP(subtype = c("enduse", "carrier"), feOnly = FALSE)
}
\arguments{
\item{subtype}{specifies share}

\item{feOnly}{specifies if shares or quantities are returned}
}
\value{
MAgPIE object with historic shares
}
\description{
Shares are calculated based on energy demands in households and services from
the IEA ETP data base.
Missing shares that result from missing demand data are filled with the
average share across all regions and periods and then normalised to sum up to
one again.
Biomass is split according to GDP per Capita (see toolSplitBiomass).
}
\author{
Robin Hasse, Antoine Levesque, Hagen Tockhorn
}
