% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareIEA_EEI.R
\name{calcShareIEA_EEI}
\alias{calcShareIEA_EEI}
\title{Calculate shares of carriers or enduses in buildings energy demand from IEA EEI data}
\usage{
calcShareIEA_EEI(subtype = c("enduse", "carrier"))
}
\arguments{
\item{subtype}{Character, specifies which type of share to calculate. Must be either
"enduse" or "carrier". Default is "enduse".}
}
\value{
A data.frame with the respective shares
}
\description{
This function calculates the proportion of energy demand by carrier or enduse
in the buildings sector, based on IEA Energy Efficiency Indicators data.
}
\author{
Hagen Tockhorn
}
