% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareOdyssee.R
\name{calcShareOdyssee}
\alias{calcShareOdyssee}
\title{Share of energy carriers or end uses in historic buildings demand}
\usage{
calcShareOdyssee(
  subtype = c("enduse", "carrier", "enduse_carrier"),
  feOnly = FALSE
)
}
\arguments{
\item{subtype}{Character, dimension names, level of shares.}

\item{feOnly}{if TRUE, output is absolute FE values}
}
\value{
MAgPIE object with historic shares
}
\description{
Shares are calculated based on energy demands in households and services from
the Odyssee data base (EU member states). In case of multi-level shares, the
function gives the share w.r.t. to the last categories but for all
categories. E.g. 'enduse_carrier' gives the share of each carrier in the
demand from each end use.
Missing shares are handled as follows: If all data for a region is missing
(no demand data at all), the shares remain NA. If a specific category (e.g.,
a specific carrier or end use) has no data across all periods for a region,
its shares are filled with 0, assuming zero demand for that category.
Biomass is split according to GDP per Capita (see toolSplitBiomass).
}
\author{
Robin Hasse, Antoine Levesque, Hagen Tockhorn
}
