% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShares.R
\name{calcShares}
\alias{calcShares}
\title{Historic carrier-Enduse Shares w.r.t Final Energy}
\usage{
calcShares(subtype = "nonthermal")
}
\value{
data.frame with historic energy demands
}
\description{
Merges and transforms the calculated shares from the Datasets:
Odyssee
IEA_ETP
}
\author{
Antoine Levesque, Hagen Tockhorn
}
