% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSharesBuildingDemand.R
\name{calcSharesBuildingDemand}
\alias{calcSharesBuildingDemand}
\title{Share of energy carriers or end uses in historic buildings demand}
\usage{
calcSharesBuildingDemand(subtype = c("enduse", "carrier"))
}
\arguments{
\item{subtype}{Character, dimension names, level of shares.}
}
\value{
MAgPIE object with historic shares
}
\description{
Shares are calculated based on energy demands from the Odyssee data base for
EU member states,
}
\author{
Robin Hasse, Antoine Levesque
}
