% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSharesEU.R
\name{calcSharesEU}
\alias{calcSharesEU}
\title{Calculate historical energy end-use shares based on final energy demands}
\usage{
calcSharesEU(thermal = FALSE, endOfHistory = 2025)
}
\arguments{
\item{thermal}{A logical specifying if end-use "appliances" shall be transformed
into "refrigerators" with appropriate share}

\item{endOfHistory}{An integer specifying the upper temporal boundary for historical data}
}
\value{
data.frame with historic end-use shares relative to final energy demands
}
\description{
Merges and transforms the calculated shares from the datasets:
Odyssee 2024
IEA Energy Efficiency Indicators 2024
IEA Tracking Clean Energy Progress
IEA World Energy Outlook 2014
}
\details{
End-use shares are extrapolated using a log-linear regression on the existing
data points while values in-between existing data are interpolated linearly.

In the thermal case, the enduse "appliances" is transformed to "refrigerators"
using the region-specific refrigerator share used in EDGE-B. Higher resolved
data was not available.
}
\author{
Hagen Tockhorn, Robin Hasse
}
