% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareTCEP.R
\name{calcShareTCEP}
\alias{calcShareTCEP}
\title{Calculate Enduse Shares from TCEP data base}
\usage{
calcShareTCEP()
}
\value{
magpie object
}
\description{
Calculate Enduse Shares from TCEP data base
}
\author{
Hagen Tockhorn
}
