% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareWEO.R
\name{calcShareWEO}
\alias{calcShareWEO}
\title{Calculate end use shares from WEO 2015}
\usage{
calcShareWEO()
}
\description{
Table 10.2 in the WEO 2015 report lists the chare of regional enduses in the
global final energy demand. Here, we calculate the share of enduses in the
final energ demand for buildings in each region.
}
\author{
Robin Hasse
}
