% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTCEP.R
\name{calcTCEP}
\alias{calcTCEP}
\title{IEA's Tracking clean energy progress 2014}
\usage{
calcTCEP(subtype = "enduse")
}
\arguments{
\item{subtype}{character, type of data}
}
\description{
IEA's Tracking clean energy progress 2014
}
\author{
Robin Hasse
}
