% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUEdemand.R
\name{calcUEdemand}
\alias{calcUEdemand}
\title{Calculate useful energy demand for space heating in buildings}
\usage{
calcUEdemand(granularity = NULL)
}
\arguments{
\item{granularity}{character, name of BRICK granularity}
}
\description{
Calculate useful energy demand for space heating in buildings
}
\author{
Robin Hasse
}
