% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUValueSource.R
\name{calcUValueSource}
\alias{calcUValueSource}
\title{Passes U-values read from source}
\usage{
calcUValueSource(subtype = c("EUBuildingsObservatory", "ETSAP"))
}
\arguments{
\item{subtype}{source of U-values}
}
\description{
Passes U-values read from source
}
\author{
Falk Benke, Robin Hasse
}
