% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWeightFeBuildings.R
\name{calcWeightFeBuildings}
\alias{calcWeightFeBuildings}
\title{calculate FE demand in buildings}
\usage{
calcWeightFeBuildings()
}
\description{
Simple estimate of total FE demand in different building types
}
\details{
Should only be used as weights not actual demand. The disaggregation to
building types is too simplistic.
}
\author{
Robin Hasse
}
