% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{cfac}
\alias{cfac}
\title{Calculate counterfactuals for solar radiation, wind speed, specific humidity
and near-surface temperature as function of raster data on near-surface temperature.}
\usage{
cfac(t, type, params = NULL)
}
\arguments{
\item{t}{raster data on near-surface atmospherical temperature}

\item{type}{considered climate variable}

\item{params}{regression parameters as vector or raster object}
}
\value{
counterfactuals for respective climate variable
}
\description{
The expected value of the respective climate variable (except temperature) is
calculated from parameters taken from a preceding linear regression done in
calcBAITpars where the respective variable is correlated with the near-surface
atmospherical temperature.
If no cell-resoluted parameters are given, the globally-meaned parameters from
Staffel et. all 2023 are taken (see https://doi.org/10.1038/s41560-023-01341-5).
}
