% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{checkDates}
\alias{checkDates}
\title{Check if time period of BAIT input data (rsds, sfc, huss) is congruent with
near-surface temperature data (tas).}
\usage{
checkDates(baitInput, tasData)
}
\arguments{
\item{baitInput}{list of raster data encompassing different climate variables}

\item{tasData}{raster data on near-surface atmosphere temperature}
}
\value{
baitInput with congruent time periods w.r.t. tasData
}
\description{
Check if time period of BAIT input data (rsds, sfc, huss) is congruent with
near-surface temperature data (tas).
}
