% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{compBAIT}
\alias{compBAIT}
\title{Calculate bias-adjusted internal temperature (BAIT)}
\usage{
compBAIT(baitInput, tasData, weight = NULL, params = NULL)
}
\arguments{
\item{baitInput}{named list containing rsds, sfcwind, huss climate data}

\item{tasData}{raster data on near-surface atmospherical temperature}

\item{weight}{named list with weights}

\item{params}{optional named list with regression parameters from calcBAITpars}
}
\value{
raster object with BAIT values
}
\description{
BAIT is calculated from raster data on near-surface atmospherical temperature
(tas), surface downdwelling shortwave radiation (rsds), near-surface wind speed
(sfcwind) and near-surface specific humidity (huss). The latter three climate
parameters are incorporated in the calculation of BAIT as the difference from
their real value to the their expected value w.r.t. the near-surface temperature
(see \code{\link{cfac}}). These are then incorporated in a weighted sum to
account for the respective influence of each climate parameter on BAIT.
The raster data containing BAIT is smoothed to account for the
buildings' thermal inertia (see \code{\link{smooth}}) and blended with the
near-surface temperature (see \code{\link{blend}}).
}
