% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{compStackHDDCDD}
\alias{compStackHDDCDD}
\title{Calculate country-wise population-weighted HDD/CDD values}
\usage{
compStackHDDCDD(
  ftas,
  tlim,
  countries,
  pop,
  factors,
  bait,
  frsds = NULL,
  fsfc = NULL,
  fhuss = NULL,
  wBAIT = NULL,
  params = NULL,
  rasDir = NULL
)
}
\arguments{
\item{ftas}{file name of data on near-surface atmospherical temperature}

\item{tlim}{named list of limit temperature sequences for \code{HDD} and \code{CDD}}

\item{countries}{raster opbject defining (regional) aggregation boundaries}

\item{pop}{raster object containing population data}

\item{factors}{data frame with degree day values for \code{temp/tlim} combination}

\item{bait}{boolean, BAIT is used as ambient temperature}

\item{frsds}{file name of data on surface downdwelling shortwave radiation (optional)}

\item{fsfc}{file name of data on near-surface wind speed (optional)}

\item{fhuss}{file name of data on near-surface specific humidity (optional)}

\item{wBAIT}{named list containing BAIT weights (optional)}

\item{params}{raster object containing regression parameters from \code{calcBAITpars} (optional)}

\item{rasDir}{absolute path to directory for saving raster files}
}
\description{
This function calculates country-wise population-weighted HDD/CDD values for
raw ambient temperature or bias-adjusted internal temperature for a given set
of limit temperatures from raster data on (various) climate variables.
}
\details{
For further processing, raster objects containing degree day data are written
for an interval of ten years.
}
