% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAssotermica.R
\name{convertAssotermica}
\alias{convertAssotermica}
\title{Convert heating systems sales in Italy from Assotermica}
\usage{
convertAssotermica(x)
}
\arguments{
\item{x}{MagPIE object with Assotermica data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Convert heating systems sales in Italy from Assotermica
}
\author{
Robin Hasse
}
