% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBDH.R
\name{convertBDH}
\alias{convertBDH}
\title{Convert German heating system data from BDH}
\usage{
convertBDH(x)
}
\arguments{
\item{x}{MagPIE object with BDH data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Convert German heating system data from BDH
}
\author{
Robin Hasse
}
