% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBMK.R
\name{convertBMK}
\alias{convertBMK}
\title{Convert heating systems sales in Austria from BMK}
\usage{
convertBMK(x)
}
\arguments{
\item{x}{MagPIE object with BMK data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Convert heating systems sales in Austria from BMK
}
\author{
Robin Hasse
}
