% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCensusHub.R
\name{convertCensusHub}
\alias{convertCensusHub}
\title{convertCensusHub}
\usage{
convertCensusHub(x, subtype)
}
\arguments{
\item{x}{raw data}

\item{subtype}{Eurostat code of data set}
}
\value{
MAgPie object with converted data
}
\description{
Convert data from Census Hub
}
\author{
Robin Hasse
}
