% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDeetman2020.R
\name{convertDeetman2020}
\alias{convertDeetman2020}
\title{Convert Weibull lifetime distribution parameters for buildings}
\usage{
convertDeetman2020(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with data from Deetman et al. 2020}

\item{subtype}{character, building subsector (either 'residential' or
'commercial')}
}
\description{
Convert Weibull lifetime distribution parameters for buildings
}
\author{
Robin Hasse
}
