% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDestatis.R
\name{convertDestatis}
\alias{convertDestatis}
\title{Convert data from the Federal Statistical Office of Germany}
\usage{
convertDestatis(x, subtype)
}
\arguments{
\item{x}{MagaPie object}

\item{subtype}{character, data code}
}
\value{
MagPIE object
}
\description{
Convert data from the Federal Statistical Office of Germany
}
\author{
Robin Hasse
}
