% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEEAfloorspace.R
\name{convertEEAfloorspace}
\alias{convertEEAfloorspace}
\title{convert EEAfloorpace}
\usage{
convertEEAfloorspace(x)
}
\arguments{
\item{x}{MAgPIE object with floor space data}
}
\value{
floor national floor space in million m2
}
\description{
Convert historic floor space data from the European Environment Agency.
}
\author{
Robin Krekeler
}
