% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEHI.R
\name{convertEHI}
\alias{convertEHI}
\title{Convert EHI Heating Market Report}
\usage{
convertEHI(x)
}
\arguments{
\item{x}{MagPIE object with raw data}
}
\value{
MagPIE object with country-level data
}
\description{
Convert EHI Heating Market Report
}
\author{
Robin Hasse
}
