% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEHPA.R
\name{convertEHPA}
\alias{convertEHPA}
\title{Correct heat pump sales data from EHPA}
\usage{
convertEHPA(x)
}
\arguments{
\item{x}{MagPIE object with EHPA data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Correct heat pump sales data from EHPA
}
\author{
Robin Hasse
}
