% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEnergiforsk2016.R
\name{convertEnergiforsk2016}
\alias{convertEnergiforsk2016}
\title{Convert European district heating prices}
\usage{
convertEnergiforsk2016(x)
}
\arguments{
\item{x}{MAgPIE object with original district heating prices}
}
\description{
We should consider prper extrapolation
}
\author{
Robin Hasse
}
