% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEUBuildingsDB.R
\name{convertEUBuildingsDB}
\alias{convertEUBuildingsDB}
\title{convertEUBuildingsDB}
\usage{
convertEUBuildingsDB(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with data from EU Buildings Database}

\item{subtype}{character <category>.<variable>}
}
\value{
clean MAgPIE object with
}
\description{
Convert to common units.
}
\author{
Robin Krekeler
}
