% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurObservER.R
\name{convertEurObservER}
\alias{convertEurObservER}
\title{convertEurObservER}
\usage{
convertEurObservER(x, subtype)
}
\arguments{
\item{x}{raw data}

\item{subtype}{Eurostat code of data set}
}
\value{
MAgPie object with converted data
}
\description{
Each report of the Heat Pumps Barometer covers the two years before. There
numerous cases, where reports contradict in overlapping values or there
are huge jumps (probably due to changes in the accounting). This function
creates smooth trajectories back-casting from the most recent data. If data
overlaps, earlier data is scaled to match more recent data. In case of gaps,
the trajectories are back-casted maintaining relative growth within each
report. Growth rates between reports are interpolated. The Method yields
mostly smooth trajectories always hitting the mos recent reported numbers.
}
\author{
Robin Hasse
}
