% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEuropeanCommissionRenovation.R
\name{convertEuropeanCommissionRenovation}
\alias{convertEuropeanCommissionRenovation}
\title{convertEuropeanCommissionRenovation}
\usage{
convertEuropeanCommissionRenovation(x)
}
\arguments{
\item{x}{MAgPIE object with data from EC report}
}
\value{
clean MAgPIE object with data from EU Buildings Database
}
\description{
Drop EU28 aggregate and rename regions to ISO3.
}
\author{
Robin Hasse
}
