% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostatBuildings.R
\name{convertEurostatBuildings}
\alias{convertEurostatBuildings}
\title{convertEurostatBuildings}
\usage{
convertEurostatBuildings(x, subtype)
}
\arguments{
\item{x}{raw data}

\item{subtype}{Eurostat code of data set}
}
\value{
MAgPie object with converted data
}
\description{
Convert buildings-related data from Eurostat.
}
\author{
Robin Hasse
}
