% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFEGECA.R
\name{convertFEGECA}
\alias{convertFEGECA}
\title{Convert heating systems sales in Spain from FEGECA}
\usage{
convertFEGECA(x)
}
\arguments{
\item{x}{MagPIE object with Assotermica data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Convert heating systems sales in Spain from FEGECA
}
\author{
Robin Hasse
}
