% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGDL.R
\name{convertGDL}
\alias{convertGDL}
\title{convertGDL}
\usage{
convertGDL(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with data from EU Buildings Database}

\item{subtype}{\if{html}{\out{<database>}}.\if{html}{\out{<variable>}}}
}
\value{
clean MAgPIE object with data from EU Buildings Database
}
\description{
Drop subnational data but keep other subgroups (i.e. rural/urban, income
quartiles)
}
\author{
Robin Hasse
}
