% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGEM.R
\name{convertGEM}
\alias{convertGEM}
\title{Convert GEM Global Exposure Model}
\usage{
convertGEM(x)
}
\arguments{
\item{x}{MAgPIE object from readGEM}
}
\value{
MAgPIE object
}
\description{
The GEM exposure data are already reported on ISO3 country level, so no
disaggregation is required. This function only fills the remaining ISO
countries with NA to obtain a complete country set.
}
\author{
Hagen Tockhorn
}
