% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGLOBUS.R
\name{convertGLOBUS}
\alias{convertGLOBUS}
\title{Convert GLOBUS building stock}
\usage{
convertGLOBUS(x)
}
\arguments{
\item{x}{MAgPIE object from readGLOBUS}
}
\value{
MAgPIE object containing building stock per ISO country
}
\description{
Disaggregate GLOBUS building stock projections from the model's aggregate
regions to ISO countries. Population is used as the disaggregation weight.
}
\author{
Hagen Tockhorn
}
