% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHDDCDDtemp.R
\name{convertHDDCDDtemp}
\alias{convertHDDCDDtemp}
\title{Convert temporary HDD and CDD data}
\usage{
convertHDDCDDtemp(x)
}
\arguments{
\item{x}{MAgPIE object with HDD and CDD data at EDGE-B resolution}
}
\value{
HDD CDD data at country resolution
}
\description{
The temporary data already is at EDGE-B regional resolution. It is uniformly
disaggregated to country level here to comply with the madrat framework
before it is aggregated to the identical resolution in fullEDGEBUILDINGS.
}
\author{
Robin Hasse
}
