% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHotmaps.R
\name{convertHotmaps}
\alias{convertHotmaps}
\title{convert Hotmaps}
\usage{
convertHotmaps(x)
}
\arguments{
\item{x}{MagPIE object with raw building stock data}
}
\value{
MagPIE object with converted data
}
\description{
convert Hotmaps
}
\author{
Robin Hasse
}
