% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHouseholdSurveys.R
\name{convertHouseholdSurveys}
\alias{convertHouseholdSurveys}
\title{Convert household survey data on energy carrier use for cooking and lighting}
\usage{
convertHouseholdSurveys(x)
}
\arguments{
\item{x}{data.frame containing survey data}
}
\description{
Convert household survey data on energy carrier use for cooking and lighting
}
\references{
https://hdl.handle.net/10419/301069
}
\author{
Hagen Tockhorn
}
