% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_OECD_Prices.R
\name{convertIEA_OECD_Prices}
\alias{convertIEA_OECD_Prices}
\title{read energy carrier prices}
\usage{
convertIEA_OECD_Prices(x, subtype)
}
\arguments{
\item{x}{MagPIE object}

\item{subtype}{character, data set}
}
\description{
read energy carrier prices
}
\author{
Robin Hasse
}
