% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEAResidentialTechnologies.R
\name{convertIEAResidentialTechnologies}
\alias{convertIEAResidentialTechnologies}
\title{convertIEAResidentialTechnologies}
\usage{
convertIEAResidentialTechnologies(x)
}
\arguments{
\item{x}{MAgPIE object with data from IEA website}
}
\value{
clean MAgPIE object
}
\description{
Convert residential technology data from IEA
}
\author{
Robin Hasse
}
