% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertISIMIPbuildings.R
\name{convertISIMIPbuildings}
\alias{convertISIMIPbuildings}
\title{convert ISIMIP data for mredgebuildings}
\usage{
convertISIMIPbuildings(x, subtype)
}
\arguments{
\item{x}{MAgPIE object on cellular level}

\item{subtype}{filename}
}
\value{
rasterBrick object
}
\description{
convert ISIMIP data for mredgebuildings
}
\author{
Hagen Tockhorn
}
