% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLTRS.R
\name{convertLTRS}
\alias{convertLTRS}
\title{Convert Long Term Renovation Strategy}
\usage{
convertLTRS(x)
}
\arguments{
\item{x}{MagPIE object with data read from source}
}
\value{
MagPIE object
}
\description{
Convert Long Term Renovation Strategy
}
\author{
Robin Hasse
}
