% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMessageIXBuildings.R
\name{convertMessageIXBuildings}
\alias{convertMessageIXBuildings}
\title{Convert data from MessageIX-Buildings}
\usage{
convertMessageIXBuildings(x, subtype)
}
\description{
Convert data from MessageIX-Buildings
}
