% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOdyssee.R
\name{convertOdyssee}
\alias{convertOdyssee}
\title{convertOdyssee}
\usage{
convertOdyssee(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with data from Odyssee Database}

\item{subtype}{dataset version}
}
\value{
clean MAgPIE object
}
\description{
Rename regions and convert unit
}
\author{
Robin Hasse
}
