% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPRISMA_ICT.R
\name{convertPRISMA_ICT}
\alias{convertPRISMA_ICT}
\title{Convert PRISMA ICT Data to Standard Format}
\usage{
convertPRISMA_ICT(x, subtype)
}
\arguments{
\item{x}{MAgPIE object from readPRISMA_ICT}

\item{subtype}{Character string specifying data type (same as in readPRISMA_ICT)}
}
\value{
A magpie object with converted country codes and filled countries
}
\description{
Converts PRISMA ICT data to ISO country codes and fills missing countries.
Primarily used for data center count data.
}
\author{
Hagen Tockhorn
}
