% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPulsen.R
\name{convertPulsen}
\alias{convertPulsen}
\title{Convert Pulsen survey on Swedish heat pump market}
\usage{
convertPulsen(x)
}
\arguments{
\item{x}{MagPIE object with data}
}
\value{
MagPIE object with corrected region
}
\description{
Convert Pulsen survey on Swedish heat pump market
}
\author{
Robin Hasse
}
