% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSandberg.R
\name{convertSandberg}
\alias{convertSandberg}
\title{Convert building life time data from Sandberg et al. 2016}
\usage{
convertSandberg(x)
}
\arguments{
\item{x}{MagPie object}
}
\value{
MagPIE object with full region converage
}
\description{
Convert building life time data from Sandberg et al. 2016
}
\author{
Robin Hasse
}
