% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSPIUG.R
\name{convertSPIUG}
\alias{convertSPIUG}
\title{Convert heating systems stock and sales in Poland from SPIUG}
\usage{
convertSPIUG(x)
}
\arguments{
\item{x}{MagPIE object with PIUG data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Convert heating systems stock and sales in Poland from SPIUG
}
\author{
Robin Hasse
}
