% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSTURM.R
\name{convertSTURM}
\alias{convertSTURM}
\title{Convert energy efficiency values from STURM (MessageIX-Buildings)}
\usage{
convertSTURM(x)
}
\description{
Convert energy efficiency values from STURM (MessageIX-Buildings)
}
