% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTCEP.R
\name{convertTCEP}
\alias{convertTCEP}
\title{Convert TCEP data base}
\usage{
convertTCEP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with data from TCEP #nolint}

\item{subtype}{character, type of data}
}
\value{
magpie object
}
\description{
Convert TCEP data base
}
\author{
Hagen Tockhorn, Robin Hasse
}
