% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNHouseholds.R
\name{convertUNHouseholds}
\alias{convertUNHouseholds}
\title{convertUNHouseholds}
\usage{
convertUNHouseholds(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with data from EU Buildings Database}

\item{subtype}{household variable}
}
\value{
clean MAgPIE object with
}
\description{
Clean up UN data on household size and composition
}
\author{
Robin Krekeler
}
