% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUniclima.R
\name{convertUniclima}
\alias{convertUniclima}
\title{Convert heating systems sales in France from Uniclima}
\usage{
convertUniclima(x)
}
\arguments{
\item{x}{MagPIE object with PIUG data}
}
\value{
MagPIE object with ISO 3 country codes
}
\description{
Convert heating systems sales in France from Uniclima
}
\author{
Robin Hasse
}
