% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVHK.R
\name{convertVHK}
\alias{convertVHK}
\title{Convert VHK data}
\usage{
convertVHK(x, subtype)
}
\arguments{
\item{x}{MagPie object}

\item{subtype}{character, specifies study and the respective csv file to
read: \code{<study>.<file>}}
}
\value{
MagPIE object
}
\description{
Convert VHK data
}
\author{
Robin Hasse
}
