% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWEO.R
\name{convertWEO}
\alias{convertWEO}
\title{Convert WEO enduse shares w.r.t. to global final energy demand}
\usage{
convertWEO(x)
}
\arguments{
\item{x}{readWEO object}
}
\description{
Convert WEO enduse shares w.r.t. to global final energy demand
}
\author{
Hagen Tockhorn
}
