% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResIRF.R
\docType{data}
\name{demand}
\alias{demand}
\title{Read input data of Res-IRF}
\format{
An object of class \code{numeric} of length 7.
}
\source{
https://doi.org/10.1016/j.eneco.2011.07.010

https://github.com/CIRED/Res-IRF/tree/v3.1/project/input
}
\usage{
demand
}
\arguments{
\item{subtype}{character, type of input data}
}
\description{
Input data of the residential sector model of the IMACLIM team.
}
\details{
The scope of the model is France.
}
\author{
Robin Hasse
}
\keyword{datasets}
