% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDisaggregate.R
\name{.disaggregate}
\alias{.disaggregate}
\title{Disaggregate energy demand within an aggregated region}
\usage{
.disaggregate(subset, key)
}
\arguments{
\item{subset}{data frame for one period and aggregated region}

\item{key}{named vector with specification of the subset group (not used)}
}
\description{
Disaggregate regional energy demand per carrier by end use while meeting the
end use shares in the aggregated region.
}
\details{
The function first tries to find a solution that satisfies both the regional
carrier and the overall end use constraints. If there is no solution, another
optimisation is run that tries to also minimise deviations from the end use
quantities but removes them from the constraints.
}
