% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dot2Dash}
\alias{dot2Dash}
\title{Replace dots by dashes}
\usage{
dot2Dash(x, rev = FALSE)
}
\arguments{
\item{x}{character vector}

\item{rev}{logical, if TRUE, replace dashes by dots}
}
\value{
character vector with replaced symbols
}
\description{
Replace dots by dashes
}
\author{
Robin Hasse
}
