% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur2usd.R
\name{eur2usd}
\alias{eur2usd}
\title{Exchange rate between EUR and USD}
\usage{
eur2usd(x = 1, year = 2020, iso3c = "DEU", ...)
}
\arguments{
\item{x}{value in EUR}

\item{year}{integer, reference year}

\item{iso3c}{region}

\item{...}{further arguments to \link[GDPuc:convertGDP]{GDPuc::convertSingle}}
}
\value{
value in constant 2017 US$MER
}
\description{
Convert EUR in given year to internal unit
}
\author{
Robin Hasse
}
