% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspacePast.R
\name{extrapolateMissingPeriods}
\alias{extrapolateMissingPeriods}
\title{Extrapolate missing values beyond existing periods}
\usage{
extrapolateMissingPeriods(chunk, key, slopeOfLast = 5)
}
\arguments{
\item{chunk}{grouped data.frame}

\item{key}{column holding values for extrapolation}

\item{slopeOfLast}{number of values for boundary regression}
}
\value{
data.frame with extrapolated column \code{key}
}
\description{
Extrapolate missing values beyond existing periods
}
