% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullVALIDATIONBUILDINGS.R
\name{fullVALIDATIONBUILDINGS}
\alias{fullVALIDATIONBUILDINGS}
\title{create historical mif for buildings}
\usage{
fullVALIDATIONBUILDINGS(rev = 0)
}
\arguments{
\item{rev}{Unused parameter, but required by \code{madrat}.}
}
\description{
create historical mif for buildings
}
\author{
Robin Hasse
}
