% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFE_by_EUEC.R
\name{getFEbyEUEC}
\alias{getFEbyEUEC}
\title{Historic Final Energy Demand by Carrier and Enduse}
\usage{
getFEbyEUEC()
}
\value{
data.frame with historic energy demands
}
\description{
Takes the historic final energy demand by carriers from IEA and disaggregates
it into different end uses.
}
\author{
Hagen Tockhorn
}
