% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_all}
\alias{join_all}
\title{Join by all common columns}
\usage{
join_all(
  x,
  y,
  by = NULL,
  .direction = "left",
  exclude = "value",
  silent = TRUE,
  ...
)
}
\arguments{
\item{x, y}{A pair of data frames}

\item{by}{character vector of columns to join by beyond the commonly named}

\item{.direction}{type of dplyr join, 'left' by default}

\item{exclude}{character vector with column names that should be excluded
from the search for join columns}

\item{silent}{boolean, whether to print type of joint the join columns}

\item{...}{additional arguments passed to join function}
}
\description{
Additional columns to join by that have different names can be give. The kind
of joint can also be specified.
}
\author{
Robin Hasse
}
