% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspacePast.R
\name{makeFloorspaceProjection}
\alias{makeFloorspaceProjection}
\title{Predict missing historic floorspace data}
\usage{
makeFloorspaceProjection(df, gdppop, dens, endOfHistory, periodBegin)
}
\arguments{
\item{df}{data.frame containing floorspace per capita}

\item{gdppop}{data.frame containing gdp per capita}

\item{dens}{data.frame containing population density}

\item{endOfHistory}{upper temporal boundary of historic data}

\item{periodBegin}{lower temporal boundary of historic data}

}
\value{
floorspace per capita with filled missing entries
}
\description{
A linear regression is performed to establish a general relationship between
floorspace per capita and gdp per capita and population density.
Predicted values are corrected w.r.t. to historical data where available,
otherwise the prediction is chosen.
}
