% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mredgebuildings-package.R
\docType{package}
\name{mredgebuildings-package}
\alias{mredgebuildings}
\alias{mredgebuildings-package}
\title{mredgebuildings: Prepare data to be used by the EDGE-Buildings model}
\description{
Prepare data to be used by the EDGE-Buildings model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mredgebuildings}
}

}
\author{
\strong{Maintainer}: Robin Hasse \email{robin.hasse@pik-potsdam.de} (\href{https://orcid.org/0000-0003-1818-3186}{ORCID})

Authors:
\itemize{
  \item Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}
  \item Antoine Levesque
  \item Hagen Tockhorn
}

}
