% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSILC.R
\name{prepareSILC}
\alias{prepareSILC}
\title{Prepare EU-SILC data.}
\usage{
prepareSILC(subtype, vars = NULL, dims = NULL, weight = "DB090")
}
\arguments{
\item{subtype}{Character, type of data.}

\item{variable}{Character vector with variable codes to include.}
}
\description{
Read selected micro data of the EU-SILC survey and aggregate it.
}
\details{
The aggregation should serve the needs of later processing but suppress the
identification of respondents.
}
\author{
Robin Hasse
}
