% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{prepBaitInput}
\alias{prepBaitInput}
\title{Read in necessary climate data to calculate BAIT or calculate mean values of
said climate data to fill missing data in case of temporal mismatch between
near-surface atmospherical temperature and other considered climate data.}
\usage{
prepBaitInput(
  frsds = NULL,
  fsfc = NULL,
  fhuss = NULL,
  baitInput = NULL,
  fillWithMean = FALSE
)
}
\arguments{
\item{frsds}{file path to raster data on surface downdwelling shortwave radiation}

\item{fsfc}{file path to raster data on near-surface wind speed}

\item{fhuss}{file path to raster data on near-surface specific humidity}

\item{baitInput}{named list of climate data}

\item{fillWithMean}{boolean, only mean is calculated and returned}
}
\value{
named list with read-in or meaned climate data
}
\description{
Read in necessary climate data to calculate BAIT or calculate mean values of
said climate data to fill missing data in case of temporal mismatch between
near-surface atmospherical temperature and other considered climate data.
}
