% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAssotermica.R
\name{readAssotermica}
\alias{readAssotermica}
\title{Read heating systems sales in Italy from Assotermica}
\usage{
readAssotermica(subtype = NULL)
}
\arguments{
\item{subtype}{Character, report to read. if NULL, the most recent report is
used for every data point.}
}
\value{
magpie object
}
\description{
Read heating systems sales in Italy from Assotermica
}
\author{
Robin Hasse
}
