% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBDH.R
\name{readBDH}
\alias{readBDH}
\title{Read German heating system data from BDH}
\usage{
readBDH()
}
\value{
MagPIE object with yearly sales of heating systems in Germany
}
\description{
Read German heating system data from BDH
}
\author{
Robin Hasse
}
