% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBEAM2.R
\name{readBEAM2}
\alias{readBEAM2}
\title{Read buildings characteristics data from the Built Environment Analysis Model (BEAM^2)}
\usage{
readBEAM2()
}
\description{
Read buildings characteristics data from the Built Environment Analysis Model (BEAM^2)
}
\author{
Hagen Tockhorn
}
