% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBMK.R
\name{readBMK}
\alias{readBMK}
\title{Read heating systems sales in Austria from BMK}
\usage{
readBMK()
}
\value{
MagPIE object with yearly sales of heating systems in Germany
}
\description{
Read heating systems sales in Austria from BMK
}
\author{
Robin Hasse
}
