% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCensusHub.R
\name{readCensusHub}
\alias{readCensusHub}
\title{Read data from the European Census Hub}
\source{
https://ec.europa.eu/CensusHub2/
}
\usage{
readCensusHub(subtype)
}
\arguments{
\item{subtype}{census subset}
}
\value{
MAgPIE object with data
}
\description{
The Census Hub of the European Statistical System provides census data from
2011 on population and housing of EU member states. 2021 census data is
announced to be published there too.
\itemize{
\item typeVintage: number of dwellings of different construction periods and
building types (single, two, multiple dwellings). Show data on: dwellings,
location: nations, Topics: Type of building, Period of construction
(select all) -> CSV (Separator: Comma)
}
}
\author{
Robin Hasse
}
