% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDaioglou.R
\name{readDaioglou}
\alias{readDaioglou}
\title{Household and demand data from Daioglou et al. 2012}
\usage{
readDaioglou(subtype = "households.specific floor space")
}
\arguments{
\item{subtype}{<dataFile>.<variable>}
}
\description{
Compilation of survey data. Quintile 0 stands for the entire population.
\itemize{
\item households: Global data on population, household size, floor space per
capita (m2/cap), population density (cap/km2), GINI and household
expenditures (PPP2005/cap). Sources are written into a dimension to allow
later cleaning based on sources
\item shares: Final energy shares of end uses in buildings. Sources are dropped.
}
}
\references{
https://doi.org/10.1016/j.energy.2011.10.044
}
\author{
Robin Hasse, Antoine Levesque
}
