% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDeetman2020.R
\name{readDeetman2020}
\alias{readDeetman2020}
\title{Read Weibull lifetime distribution parameters for buildings}
\source{
https://doi.org/10.1016/j.jclepro.2019.118658
}
\usage{
readDeetman2020(subtype = "residential")
}
\arguments{
\item{subtype}{character, building subsector (either 'residential' or
'commercial')}
}
\value{
MagPIE object with Weibull lifetime parameters for different world
regions
}
\description{
Read Weibull lifetime distribution parameters for buildings
}
\author{
Robin Hasse
}
