% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDestatis.R
\name{readDestatis}
\alias{readDestatis}
\title{Read data from the Federal Statistical Office of Germany}
\usage{
readDestatis(subtype)
}
\arguments{
\item{subtype}{character, data code
\itemize{
\item \code{"31121-0005"}: Construction work completed regarding new
buildings
}}
}
\value{
MagPIE object
}
\description{
Read data from the Federal Statistical Office of Germany
}
\author{
Robin Hasse
}
