% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readECEMF.R
\name{readECEMF}
\alias{readECEMF}
\title{Read ECEMF data}
\usage{
readECEMF(subtype = "FEPrices")
}
\arguments{
\item{subtype}{character, type of data}
}
\description{
Project data: European Climate and Energy Modelling Forum
}
\details{
\itemize{
\item FEPrices: price components w/o VAT in EUR/MWh
}
}
\author{
Robin Hasse
}
