% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEEAfloorspace.R
\name{readEEAfloorspace}
\alias{readEEAfloorspace}
\title{Read national floor space from EEA}
\usage{
readEEAfloorspace()
}
\value{
magpie object with floor space
}
\description{
Floor space for many European countries between 1990 and 2009 in thousand m2.
}
\examples{
\dontrun{ a <- readSource(type = "EEAfloorspace")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Robin Krekeler
}
