% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEHI.R
\name{readEHI}
\alias{readEHI}
\title{Read EHI Heating Market Report}
\usage{
readEHI(subtype = NULL)
}
\arguments{
\item{subtype}{Character, report to read. if NULL, the most recent report is
used for every data point.}
}
\value{
magpie object
}
\description{
Read EHI Heating Market Report
}
\author{
Robin Hasse
}
