% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEHPA.R
\name{readEHPA}
\alias{readEHPA}
\title{Read heat pump sales data from EHPA}
\usage{
readEHPA(subtype = "hpSales")
}
\arguments{
\item{subtype}{character, type of data}
}
\value{
MagPIE object with HP sales data for EU member states
}
\description{
This data was provided upon request. It is free for non-profit organisations
but we are not allowed to share it.
}
\author{
Robin Hasse
}
