% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEnergiforsk2016.R
\name{readEnergiforsk2016}
\alias{readEnergiforsk2016}
\title{Read European district heating prices}
\usage{
readEnergiforsk2016()
}
\description{
Read European district heating prices
}
\author{
Robin Hasse
}
