% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEUBuildingsDB.R
\name{readEUBuildingsDB}
\alias{readEUBuildingsDB}
\title{Read EU Buildings Database}
\usage{
readEUBuildingsDB(subtype)
}
\arguments{
\item{subtype}{character <category>.<variable>}
}
\value{
magpie object
}
\description{
Compilation of various data sources from the European Commission. Because of
the considereable size of this database, a subtype has to be provided. The
subtype represents the category of items. It is possible to append selected
variables for further filtering.
}
\examples{
\dontrun{ a <- readSource(type = "EUBuildingsDB")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Robin Krekeler
}
