% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEUBuildingsObservatory.R
\name{readEUBuildingsObservatory}
\alias{readEUBuildingsObservatory}
\title{Read in energy efficiency values of building shell in residential and
non-residential buildings by the EU Buildings Observatory (2017)}
\usage{
readEUBuildingsObservatory()
}
\description{
Read in energy efficiency values of building shell in residential and
non-residential buildings by the EU Buildings Observatory (2017)
}
