% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFEGECA.R
\name{readFEGECA}
\alias{readFEGECA}
\title{Read heating systems sales in Spain from FEGECA}
\usage{
readFEGECA(subtype = NULL)
}
\arguments{
\item{subtype}{Character, report to read. if NULL, the most recent report is
used for every data point.}
}
\value{
magpie object
}
\description{
Read heating systems sales in Spain from FEGECA
}
\author{
Robin Hasse
}
