% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDL.R
\name{readGDL}
\alias{readGDL}
\title{Read Global Data Lab}
\source{
https://globaldatalab.org/
}
\usage{
readGDL(subtype)
}
\arguments{
\item{subtype}{\if{html}{\out{<database>}}.\if{html}{\out{<variable>}}}
}
\value{
magpie object
}
\description{
Different databases each comprising the harmonised results of
regional/national surveys. Available subtypes are
\itemize{
\item SHDI: Subnational Human development Database
\item AreaDatabase: Subnational development Indicators (low/medium income countries)
}
}
\examples{
\dontrun{ a <- readSource(type = "GDL")
}

}
\references{
https://doi.org/10.1038/sdata.2019.38

https://globaldatalab.org/asset/286/Smits\%20GDL\%20Working\%
20Paper\%2016-101\%20v360.pdf
}
\seealso{
\code{\link{readSource}}
}
\author{
Robin Hasse
}
