% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEM.R
\name{readGEM}
\alias{readGEM}
\title{Read GEM Global Exposure Model}
\usage{
readGEM()
}
\value{
MAgPIE object
}
\description{
Building-stock exposure figures from the GEM (Global Earthquake Model) global
exposure model.
}
\author{
Hagen Tockhorn
}
