% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGLOBUS.R
\name{readGLOBUS}
\alias{readGLOBUS}
\title{Read GLOBUS building stock}
\usage{
readGLOBUS()
}
\value{
MAgPIE object containing building stock data
}
\description{
Building stock projections for residential, non-residential and total building stock in
million m2 under the NR (no renovation) scenario from the GLOBUS building turnover model.
}
\references{
https://doi.org/10.1016/j.ynexs.2024.100019
}
\author{
Hagen Tockhorn
}
