% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHDDCDD.R
\name{readHDDCDD}
\alias{readHDDCDD}
\title{Read historical and projected degree-day data}
\usage{
readHDDCDD(subtype = "ssp2")
}
\arguments{
\item{subtype}{specifies SSP scenarios}
}
\description{
Read historical and projected degree-day data
}
\author{
Hagen Tockhorn
}
