% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHDDCDDtemp.R
\name{readHDDCDDtemp}
\alias{readHDDCDDtemp}
\title{Temporary reading of HDD and CDD data}
\usage{
readHDDCDDtemp()
}
\description{
Heating and cooling degree days calculated with \code{mredgebuildings::calcHDDCDD}.
Calculating this data currently requires manual help. We therefore read the
result as a source here. This becomes obsolete once the primary calculation
is stable.
}
\author{
Robin Hasse
}
