% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHotmaps.R
\name{readHotmaps}
\alias{readHotmaps}
\title{read EU28 building stock from Hotmaps}
\usage{
readHotmaps()
}
\description{
Consistent building stock of EU28 residential and commercial sector. The data
presents a snapshot, I guess in 2016 but this is not clear to e yet.
}
\author{
Robin Hasse
}
