% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHouseholdSurveys.R
\name{readHouseholdSurveys}
\alias{readHouseholdSurveys}
\title{Read household survey data on energy carrier use for cooking and lighting}
\usage{
readHouseholdSurveys(subtype = c("cooking", "lighting"))
}
\arguments{
\item{subtype}{Character specifying end-use}
}
\description{
Read household survey data on energy carrier use for cooking and lighting
}
\references{
https://hdl.handle.net/10419/301069
}
\author{
Hagen Tockhorn
}
