% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_acOwnership.R
\name{readIEA_acOwnership}
\alias{readIEA_acOwnership}
\title{Read AC ownership rates per household from IEA Sources}
\usage{
readIEA_acOwnership()
}
\description{
Read AC ownership rates per household from IEA Sources
}
\author{
Hagen Tockhorn
}
