% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_coolingEfficiencies.R
\name{readIEA_coolingEfficiencies}
\alias{readIEA_coolingEfficiencies}
\title{Read IEA Global Weighted Average of Air Conditioning Efficiencies}
\usage{
readIEA_coolingEfficiencies()
}
\description{
Read IEA Global Weighted Average of Air Conditioning Efficiencies
}
\author{
Hagen Tockhorn
}
