% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_OECD_Prices.R
\name{readIEA_OECD_Prices}
\alias{readIEA_OECD_Prices}
\title{read energy carrier prices}
\usage{
readIEA_OECD_Prices(subtype)
}
\arguments{
\item{subtype}{character, data set}
}
\description{
read energy carrier prices
}
\author{
Robin Hasse
}
