% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEAfloorspace.R
\name{readIEAfloorspace}
\alias{readIEAfloorspace}
\title{Floor space by subsectors}
\source{
https://iea.blob.core.windows.net/assets/416e9555-67f9-49ed-95ed-8afccd71b433/
Tracking_clean_energy_progress_2014.pdf
}
\usage{
readIEAfloorspace()
}
\value{
MAgPIE object with floor space
}
\description{
Residential and commercial floor space for big world regions from the IEA
TCEP report 2014 in billion m2.
}
\examples{
\dontrun{ a <- readSource(type = "IEAfloorspace")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Robin Krekeler
}
