% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEAResidentialTechnologies.R
\name{readIEAResidentialTechnologies}
\alias{readIEAResidentialTechnologies}
\title{Read residential technology data from IEA}
\source{
https://www.iea.org/articles/are-renewable-heating-options-cost-
competitive-with-fossil-fuels-in-the-residential-sector
}
\usage{
readIEAResidentialTechnologies()
}
\description{
techno-economic data of residential heating technologies
}
\author{
Robin Hasse
}
