% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIPbuildings.R
\name{readISIMIPbuildings}
\alias{readISIMIPbuildings}
\title{Read relevant ISIMIP data for mredgebuildings}
\usage{
readISIMIPbuildings(subtype)
}
\arguments{
\item{subtype}{filename}
}
\description{
Read relevant ISIMIP data for mredgebuildings
}
\author{
Hagen Tockhorn
}
