% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLTRS.R
\name{readLTRS}
\alias{readLTRS}
\title{Read Long Term Renovation Strategy}
\usage{
readLTRS()
}
\value{
MagPIE object
}
\description{
In the LTRS, EU member states have to describe their building stock.
Unfortunately, this isn't standardised in any way so it is hard to make use
of this data. So far, we only use data from two figures in the German LTRS.
}
\author{
Robin Hasse
}
