% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMessageIXBuildings.R
\name{readMessageIXBuildings}
\alias{readMessageIXBuildings}
\title{Read data from MessageIX-Buildings}
\usage{
readMessageIXBuildings(subtype = "uvalue")
}
\description{
Read data from MessageIX-Buildings
}
