% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPFUDB.R
\name{readPFUDB}
\alias{readPFUDB}
\title{Read Data from Primary Fuel and Useful Energy Data Base (PFUDB)}
\usage{
readPFUDB()
}
\value{
magpie object
}
\description{
This was adapted from EDGE function 'getPFUDB.R'.
}
\author{
Hagen Tockhorn
}
