% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPRISMA_ICT.R
\name{readPRISMA_ICT}
\alias{readPRISMA_ICT}
\title{Read PRISMA ICT Data}
\usage{
readPRISMA_ICT(subtype)
}
\arguments{
\item{subtype}{Character string specifying data type}
}
\value{
A magpie object containing the requested ICT/data center data
}
\description{
Reads ICT and data center electricity demand data from PRISMA project Excel file.
Supports different regional resolutions and data center counts.
}
\author{
Hagen Tockhorn
}
