% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPulsen.R
\name{readPulsen}
\alias{readPulsen}
\title{Read Pulsen survey on Swedish heat pump market}
\source{
https://skvp.se/skvpold/statistik/pulsen
}
\usage{
readPulsen()
}
\value{
MagPIE object
}
\description{
Read Pulsen survey on Swedish heat pump market
}
\author{
Robin Hasse
}
