% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSandberg.R
\name{readSandberg}
\alias{readSandberg}
\title{Read building life time data from Sandberg et al. 2016}
\source{
https://doi.org/10.1016/j.enbuild.2016.05.100
}
\usage{
readSandberg()
}
\value{
MagPIE object with life time parameters of buildings
}
\description{
Read building life time data from Sandberg et al. 2016
}
\author{
Robin Hasse
}
