% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSPIUG.R
\name{readSPIUG}
\alias{readSPIUG}
\title{Read heating systems stock and sales in Poland from SPIUG}
\usage{
readSPIUG()
}
\value{
magpie object
}
\description{
Read heating systems stock and sales in Poland from SPIUG
}
\author{
Robin Hasse
}
