% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSTURM.R
\name{readSTURM}
\alias{readSTURM}
\title{Read in energy efficiency values from STURM (MessageIX-Buildings)}
\usage{
readSTURM()
}
\description{
Read in energy efficiency values from STURM (MessageIX-Buildings)
}
