% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCEP.R
\name{readTCEP}
\alias{readTCEP}
\title{Read TCEP data base}
\usage{
readTCEP(subtype = "enduse")
}
\arguments{
\item{subtype}{character, type of data}
}
\value{
magpie object
}
\description{
Read TCEP data base
}
\author{
Hagen Tockhorn, Robin Hasse
}
