% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNHouseholds.R
\name{readUNHouseholds}
\alias{readUNHouseholds}
\title{Read UN data on Household Size & Composition}
\source{
https://www.un.org/development/desa/pd/data/household-size-and-composition
}
\usage{
readUNHouseholds(subtype)
}
\arguments{
\item{subtype}{household variable}
}
\value{
magpie object
}
\description{
Compilation of indicators on household size and membership composition. The
estimates are based on 814 unique data points from 172 countries or areas,
representing approximately 97 per cent of the world’s population in 2019,
collected between 1960 to 2018.
}
\examples{
\dontrun{ a <- readSource(type = "UNHouseholds")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Robin Hasse
}
