% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUniclima.R
\name{readUniclima}
\alias{readUniclima}
\title{Read heating systems sales in France from Uniclima}
\usage{
readUniclima()
}
\value{
magpie object
}
\description{
Read heating systems sales in France from Uniclima
}
\author{
Robin Hasse
}
