% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVHK.R
\name{readVHK}
\alias{readVHK}
\title{Read VHK data}
\usage{
readVHK(subtype)
}
\arguments{
\item{subtype}{character, specifies study and the respective csv file to
read: \code{<study>.<file>}}
}
\value{
MagPIE object
}
\description{
multiple EU studies mostly on different categories of appliances
}
\author{
Robin Hasse
}
