% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWEO.R
\name{readWEO}
\alias{readWEO}
\title{Read WEO enduse shares w.r.t. to global final energy demand}
\usage{
readWEO(subtype)
}
\arguments{
\item{subtype}{variable to define data subset ("Buildings", "Transport", "Industry")}
}
\description{
Read WEO enduse shares w.r.t. to global final energy demand
}
\author{
Hagen Tockhorn
}
