% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{smooth}
\alias{smooth}
\title{Smooth data over preceding two days}
\usage{
smooth(r, weight)
}
\arguments{
\item{r}{raster data to be smoothed}

\item{weight}{named list with smoothing parameter sig}
}
\value{
smoothed raster data
}
\description{
Smooth data over preceding two days
}
