% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalcShares.R
\name{toolCalcShares}
\alias{toolCalcShares}
\title{Calculate Shares}
\usage{
toolCalcShares(df, colShare)
}
\arguments{
\item{df}{data.frame}

\item{colShare}{column to be grouped over}
}
\value{
data.frame with calculated shares for colShare
}
\description{
Calculate Shares
}
\author{
Hagen Tockhorn
}
