% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCompleteRenFlows.R
\name{toolCompleteRenFlows}
\alias{toolCompleteRenFlows}
\title{Complete Matrix of flows from old to new to cover given items}
\usage{
toolCompleteRenFlows(x, from, to)
}
\arguments{
\item{x}{MagPIE object, renovation flows to complete}

\item{from}{MagPIE object with origin quantities}

\item{to}{MagPIE object with destination quantities}
}
\value{
MagPIE object
}
\description{
Extrapolate from data on selected technologies to all technologies required
}
\details{
We assume that there is a high status quo probability that people just do
identical replacement. If they don't, they chose the system with
probabilities proportional to the general shares.
}
\author{
Robin Hasse
}
