% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCountryFillAvg.R
\name{toolCountryFillAvg}
\alias{toolCountryFillAvg}
\title{Tool: CountryFillAvg}
\usage{
toolCountryFillAvg(x, ...)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes in the spatial dimension}

\item{...}{arguments passed to toolCountryFill}
}
\description{
This function expects a MAgPIE object with ISO country codes in the spatial
dimension. It applies toolCountryFill but fills with the average across
regions.
}
\author{
Robin Hasse
}
