% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDisaggregate.R
\name{toolDisaggregate}
\alias{toolDisaggregate}
\title{Disaggregates Enduse and Carrier Shares into combined enduse-carrier shares
via a quadratic optimization for an under-determined linear set of equations.}
\usage{
toolDisaggregate(
  data,
  sharesEU,
  etpEU,
  exclude = NULL,
  sharesReplace = NULL,
  correct = FALSE
)
}
\arguments{
\item{data}{data.frame containing data on carrier-specific energy consumption.
Columns 'carrier' and 'value' are required.}

\item{sharesEU}{data.frame containing enduse shares}

\item{etpEU}{data.frame containing enduse energy data}

\item{exclude}{list containing "enduse-carrier" pairs to be excluded}

\item{sharesReplace}{data.frame containing existing disaggregated carrier-enduse shares}

\item{correct}{boolean that determines whether values shall be corrected for}
}
\description{
Disaggregates Enduse and Carrier Shares into combined enduse-carrier shares
via a quadratic optimization for an under-determined linear set of equations.
}
\author{
Hagen Tockhorn
}
