% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetBrickMapping.R
\name{toolGetBrickMapping}
\alias{toolGetBrickMapping}
\title{Get BRICK aggregation mappings}
\usage{
toolGetBrickMapping(granularity)
}
\arguments{
\item{granularity}{character, name of BRICK granularity}
}
\value{
named list of brick dimension aggregation mappings
}
\description{
Provides all mappings needed to aggregate BRICK input data to a lower
granularity.
}
\author{
Robin Hasse
}
