% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSplitBiomass.R
\name{toolSplitBiomass}
\alias{toolSplitBiomass}
\title{Splits Biomass according to GDP per Capita}
\usage{
toolSplitBiomass(df, df_gdppop, min = 10000, max = 15000, varName = "biomass")
}
\arguments{
\item{df}{dataframe with carrier as 'variable'}

\item{df_gdppop}{dataframe containing the GDP per Capita}

\item{min}{lower threshold of GDP per Capita}

\item{max}{upper threshold of GDP per Capita}

\item{varName}{nomenclature of biomass}

\item{newvarName}{nomenclature of mod/trad biomass}
}
\value{
factor.data.frame
}
\description{
This is adapted from EDGE-B Function "splitBiomass.R" by Antoine Levesque.
}
\details{
Any value belonging to the carrier defined by "varName" will be split into
the default carrier names "biomod" and "biotrad" according to weights given
by "df_gdppop".
}
\author{
Hagen Tockhorn
}
