% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolUnitConversion.R
\name{toolUnitConversion}
\alias{toolUnitConversion}
\title{Convert unit}
\usage{
toolUnitConversion(x, unitConversion, removeUnit = FALSE)
}
\arguments{
\item{x}{MAgPIE object}

\item{unitConversion}{data.frame with the columns from, to, factor}

\item{removeUnit}{boolean remove unit after conversion for clean variable
names}
}
\description{
Takes MAgPIE object with a variable dimension and a unit dimension,
and converts the unit according to a unit conversion table.
}
\author{
Robin Hasse
}
