% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usd2eur.R
\name{usd2eur}
\alias{usd2eur}
\title{Exchange rate between USD and EUR}
\usage{
usd2eur(year = 2020)
}
\arguments{
\item{year}{integer, reference year}
}
\value{
MER EUR in USD in given year
}
\description{
Exchange rate between USD and EUR
}
\author{
Robin Hasse
}
