% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEbyEUEC.R
\name{calcFEbyEUEC}
\alias{calcFEbyEUEC}
\title{Historic Final Energy Demand by Carrier and Enduse}
\usage{
calcFEbyEUEC()
}
\value{
data.frame with historic energy demands
}
\description{
Takes the historic final energy demand by carriers from IEA and disaggregates
it into different end uses.
}
\details{
The disaggregation is performed such that the aggregated carrier-specific FE
values of the IEA IO regions are met as a minimum requirement. In an ideal case,
the enduse-specific FE shares are met as well.
}
\note{
For now, existing disaggregated final energy data with respect to carriers and
enduses combined is replaced in the final output. However, since the Odyssee
data is largely underestimating the real IEA FE targets, these shall serve as
lower boundaries for the disaggregation.
Such feature has been implemented in \code{toolDisaggregate} but is not running
smoothly yet.
}
\author{
Hagen Tockhorn, Robin Hasse
}
