% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspacePast.R
\name{calcFloorspacePast}
\alias{calcFloorspacePast}
\title{Historical residential floor space demand}
\usage{
calcFloorspacePast()
}
\value{
MAgPIE object with historic floor space
}
\description{
Data for many European countries is taken from EEA, India from IEA and other
countries from Daioglou et al 2012. The result does not cover all countries
and has mixed points in time depending on the region.
}
\note{
RK: In Antoine's EDGE-B, data points associated with an GDP/POP above
70000 USD/cap are dropped here to improve the later regression. This
filtering should be moved to getFloorspaceResidential where the regression is
performed. Therefore, the high-income data points are kept at this stage.
}
\author{
Robin Krekeler, Antoine Levesque
}
