% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspacePerCap.R
\name{calcFloorspacePerCap}
\alias{calcFloorspacePerCap}
\title{Residential floor space per capita}
\usage{
calcFloorspacePerCap(scenarios, granularity = NULL)
}
\arguments{
\item{scenarios}{A string (or vector of strings) designating the scenario(s) to be returned.}

\item{granularity}{character, name of BRICK granularity}
}
\value{
MagPIE object with residential floor space per capita
}
\description{
Residential floor space per capita
}
\author{
Robin Hasse
}
