% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspacePerCap.R
\name{calcFloorspacePerCap}
\alias{calcFloorspacePerCap}
\title{Residential floor space per capita}
\usage{
calcFloorspacePerCap()
}
\value{
MagPIE object with residential floor space per capita
}
\description{
Residential floor space per capita
}
\author{
Robin Hasse
}
