% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPPop.R
\name{calcGDPPop}
\alias{calcGDPPop}
\title{Get historic GDP per Capita for SSP Scenarios}
\usage{
calcGDPPop()
}
\value{
magpie object
}
\description{
Get historic GDP per Capita for SSP Scenarios
}
\author{
Hagen Tockhorn
}
