% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{calcHDDCDD}
\alias{calcHDDCDD}
\title{HDD and CDD scenario and historic data}
\usage{
calcHDDCDD(tlimit = 18)
}
\arguments{
\item{tlimit}{Temperature threshold for computing HDD and CDD: 18, 21 or 25}
}
\description{
smoothes the HDD and CDD scenario data
}
\author{
Antoine Levesque
}
