% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{calcHDDCDD}
\alias{calcHDDCDD}
\title{Calculate HDD and CDD based on outdoor/indoor temperature difference}
\usage{
calcHDDCDD(
  mappingFile,
  bait = FALSE,
  multiscen = FALSE,
  rasDir = NULL,
  cacheDir = NULL,
  fromSource = FALSE
)
}
\arguments{
\item{mappingFile}{file name of sectoral mapping containing input data file names and directories}

\item{bait}{boolean, use BAIT instead of raw temperature}

\item{multiscen}{boolean, does \code{mappingFile} cover more than one scenario?}

\item{rasDir}{absolute path to directory for saving raster files}

\item{cacheDir}{absolute path to directory for pre-calculated BAIT regression parameters}

\item{fromSource}{logical, if TRUE, read the previously calculated data that
is temporarily saved as a source}
}
\value{
magpie object of heating and cooling degree days
}
\description{
Heating and cooling degree days based on raw outside temperature
or bias-adjusted internal temperature (BAIT), driver for space heating and
cooling demand in buildings.
}
\author{
Robin Hasse, Hagen Tockhorn
}
