% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDDCDD.R
\name{calcHDDCDD}
\alias{calcHDDCDD}
\title{calculate HDD and CDD based on outdoor/indoor temperature difference}
\usage{
calcHDDCDD(mappingFile, bait = FALSE, multiscen = FALSE)
}
\arguments{
\item{mappingFile}{data.frame containing input data file names and directories}

\item{bait}{specify use of raw temperature or BAIT}

\item{multiscen}{specify if mappingFile covers more than one scenario}
}
\value{
magpie object of heating and cooling degree days
}
\description{
heating and cooling degree days based on raw outside temperature
or bais-adjusted internal temperature (BAIT), driver for space heating and
cooling demand in buildings
}
\examples{

\dontrun{
calcHDDCDD()
}

}
\author{
Robin Krekeler, Hagen Tockhorn
}
