% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeatingSystem.R
\name{calcHeatingSystem}
\alias{calcHeatingSystem}
\title{Calculate installation cost or efficiency of heating systems}
\usage{
calcHeatingSystem(subtype = c("Purchasing cost", "Efficiency"))
}
\arguments{
\item{subtype}{character, variable type (either 'Purchasing cost' or
'Efficiency')}
}
\value{
MagPIE object with capacity-specific purchasing cost or efficiency
of heating systems
}
\description{
Calculate installation cost or efficiency of heating systems
}
\author{
Robin Hasse
}
