% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_EEI.R
\name{calcIEA_EEI}
\alias{calcIEA_EEI}
\title{Process data from IEA End Uses and Efficiency Indicators Database}
\usage{
calcIEA_EEI(subtype = c("buildings"))
}
\arguments{
\item{subtype}{sector name}
}
\value{
data.frame containing enduse- and carrier-resoluted energy data
}
\description{
IEA EEI final energy data is processed and mapped w.r.t. carrier and enduse names.
}
\details{
As for the buildings sector, data for residential and commercial ("service")
buildings is aggregated and the carrier "biomass" is split into traditional
and modern biomass w.r.t. to income per capita.
}
\author{
Hagen Tockhorn
}
