% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationBuildings.R
\name{calcPopulationBuildings}
\alias{calcPopulationBuildings}
\title{Historic and projected population}
\usage{
calcPopulationBuildings()
}
\description{
Population differentiated by location (urban/rural) and building type
}
\author{
Robin Hasse
}
