% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareChina.R
\name{calcShareChina}
\alias{calcShareChina}
\title{Calculate shares from final energy end-use demand for China}
\usage{
calcShareChina()
}
\description{
The data is obtained from the 2021 IEA report "An Energy Sector Roadmap to Carbon
Neutrality in China" from Figure 3.25
}
\author{
Robin Hasse
}
