% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareETP.R
\name{calcShareETP}
\alias{calcShareETP}
\title{Share of energy carriers or end uses in historic buildings demand}
\usage{
calcShareETP(subtype = c("enduse", "carrier"))
}
\arguments{
\item{subtype}{specifies share}
}
\value{
MAgPIE object with historic shares
}
\description{
Shares are calculated based on energy demands in households and services from
the IEA ETP data base.
Missing shares that result from missing demand data are filled with the
average share across all regions and periods and then normalised to sum up to
one again. Traditional biomass use is assumed to be zero.
}
\author{
Robin Hasse, Antoine Levesque, Hagen Tockhorn
}
