% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareOdyssee.R
\name{calcShareOdyssee}
\alias{calcShareOdyssee}
\title{Share of energy carriers or end uses in historic buildings demand}
\usage{
calcShareOdyssee(subtype = c("enduse", "enduse_carrier"))
}
\value{
MAgPIE object with historic shares
}
\description{
Shares are calculated based on energy demands in households and services from
the Odyssee data base (EU member states). In case of multi-level shares, the
function gives the share w.r.t. to the last categories but for all
categories. E.g. 'enduse_carrier' gives the share of each carrier in the
demand from each end use.
Missing shares that result from missing demand data are filled with the
average share across all regions and periods and then normalised to sum up to
one again. Traditional biomass use is assumed to be zero.
}
\author{
Robin Hasse, Antoine Levesque
}
