% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUEdemand.R
\name{calcUEdemand}
\alias{calcUEdemand}
\title{Calculate useful energy demand for space heating in buildings}
\usage{
calcUEdemand()
}
\description{
Calculate useful energy demand for space heating in buildings
}
\author{
Robin Hasse
}
