% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMessageIXBuildings.R
\name{convertMessageIXBuildings}
\alias{convertMessageIXBuildings}
\title{Convert data from MessageIX-Buildings}
\usage{
convertMessageIXBuildings(x, subtype)
}
\arguments{
\item{x}{read-in data object}

\item{subtype}{specifies data base category}
}
\description{
Convert data from MessageIX-Buildings
}
