% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPFUDB.R
\name{convertPFUDB}
\alias{convertPFUDB}
\title{Converts Data from Primary Fuel and Useful Energy Data Base (PFUDB)}
\usage{
convertPFUDB(x)
}
\arguments{
\item{x}{MAgPIE object with data from PFUDB #nolint}
}
\value{
magpie object
}
\description{
This was adapted from EDGE function 'getPFUDB.R'.
}
\author{
Hagen Tockhorn
}
