% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usd2eur.R
\name{eur2usd}
\alias{eur2usd}
\title{Exchange rate between EUR and USD}
\usage{
eur2usd(year = 2020)
}
\arguments{
\item{year}{integer, reference year}
}
\value{
MER EUR in USD in given year
}
\description{
Convert EUR in given year to internal unit
}
\author{
Robin Hasse
}
