% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullBRICK.R
\name{fullBRICK}
\alias{fullBRICK}
\title{fullBRICK}
\usage{
fullBRICK(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).}
}
\description{
Compute complete input data set for BRICK
}
\author{
Robin Hasse
}
