% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDaioglou.R
\name{readDaioglou}
\alias{readDaioglou}
\title{Household data from Daioglou et al. 2012}
\usage{
readDaioglou(subtype = "specific floor space")
}
\arguments{
\item{subtype}{data type}
}
\description{
Global data on population, household size, floor space per capita (m2/cap),
population density (cap/km2), GINI and household expenditures (PPP2005/cap)
from various sources. Quintile 0 stands for the entire population.
}
\author{
Robin Krekeler, Antoine Levesque
}
