% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEHI.R
\name{readEHI}
\alias{readEHI}
\title{Read EHI Heating Market Report}
\usage{
readEHI(subtype)
}
\arguments{
\item{subtype}{\if{html}{\out{<report series>}}.\if{html}{\out{<variable>}}}
}
\value{
magpie object
}
\description{
Read EHI Heating Market Report
}
\author{
Robin Hasse
}
