% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanCommissionRenovation.R
\name{readEuropeanCommissionRenovation}
\alias{readEuropeanCommissionRenovation}
\title{EU report on historic renovation}
\source{
https://op.europa.eu/en/publication-detail/-/publication/97d6a4ca-5847-11ea-8b81-01aa75ed71a1
}
\usage{
readEuropeanCommissionRenovation(subtype = "csv")
}
\arguments{
\item{subtype}{character, source to read from: \code{"csv"} (default) or \code{"pdf"}
(requires the suggested package pdftools)}
}
\value{
MAgPIE object with renovation data
}
\description{
This function reads data on renovation from a pdf file of an EC report from
2019 (Tables 2 - 13). There is no period dimension. The data represents
averages between 2012 and 2016.
}
\examples{
\dontrun{ a <- readSource(type = "EuropeanCommissionRenovation")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Robin Hasse
}
