% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostatBuildings.R
\name{readEurostatBuildings}
\alias{readEurostatBuildings}
\title{Read Eurostat data on buildings}
\usage{
readEurostatBuildings(subtype)
}
\arguments{
\item{subtype}{Eurostat code of data set}
}
\value{
MAgPIE object with data
}
\description{
Read various data sets from Eurostat that are used in the modelling of
buildings
\itemize{
\item nrg_inf_hptc: Heat pumps - technical characteristics by technologies
\item ilc_hcmh02: Average size of dwelling by household type and degree of urbanisation
\item nrg_d_hhq: Disaggregated final energy consumption in households
}
}
\note{
see https://ec.europa.eu/eurostat/web/energy/data/energy-balances
for definitions of codes
}
\author{
Robin Hasse
}
