% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHDDCDD.R
\name{readHDDCDD}
\alias{readHDDCDD}
\title{Read HDD CDD}
\usage{
readHDDCDD(subtype = 18)
}
\arguments{
\item{subtype}{Temperature threshold for computing HDD and CDD}
}
\value{
magpie object HDD CDD
}
\description{
Read heating and cooling degree days; past and scenario data as magclass
objects
}
\examples{

\dontrun{ a <- readSource(type="HDDCDD")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Antoine Levesque
}
