% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_China.R
\name{readIEA_China}
\alias{readIEA_China}
\title{Read final energy data by energy end-use for China}
\usage{
readIEA_China()
}
\description{
The data is obtained from the 2021 IEA report "An Energy Sector Roadmap to Carbon
Neutrality in China" from Figure 3.25
}
\author{
Robin Hasse
}
