% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIPbuildings.R
\name{readISIMIPbuildings}
\alias{readISIMIPbuildings}
\title{Read relevant ISIMIP data for mredgebuildings}
\usage{
readISIMIPbuildings(subtype)
}
\arguments{
\item{subtype}{filename}
}
\description{
Relevant data such as region masks, population and relevant climate data are
read in. The relevant file is declared in the subtype with the full file name.
}
\details{
If the file name includes a suffix in the form of an integer such as
\code{_<int>.filetype}, the file is split into a single year period, e.g.
\code{<filename>_2001_2010_2.nc} will return data for the second year of the
2001-2010 period, here 2002.
}
\note{
folder structure in inputdata/sources/ISIMIPbuildings is expected to be:
country masks : var/
population :    var/scenario
other :         var/scenario/model

currently, this function only reads data from ISIMIP3b
}
\author{
Hagen Tockhorn
}
