% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOdyssee.R
\name{readOdyssee}
\alias{readOdyssee}
\title{Read Odyssee data base}
\usage{
readOdyssee(subtype = "enduseShares")
}
\arguments{
\item{subtype}{type of result (\code{enduseShares} or \code{carrierShares})}
}
\value{
data.frame with historic enduse or carrier shares
}
\description{
Read and tidy data on final energy consumption for different end uses in
residential (households) and commercial (services) buildings in Europe.
Calculates end use or carrier shares for European EDGE regions.
}
\references{
Enerdata. Odyssee-Mure database. https://www.odyssee-mure.eu/
}
\author{
Pascal Führlich, Antoine Levesque
}
