% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOdyssee.R
\name{readOdyssee}
\alias{readOdyssee}
\title{Read Odyssee data base}
\source{
https://odyssee.enerdata.net/database/
}
\usage{
readOdyssee(subtype = "households")
}
\arguments{
\item{subtype}{database category}
}
\value{
magpie object
}
\description{
Copied from website: The Odyssee database contains detailed energy
consumption by end-use and their drivers as well as energy efficiency and
CO2-related indicators. Latest available data is provided by national
representatives, such as energy agencies or statistical organization, from
all EU countries as well as Norway, Serbia, Switzerland and the United
Kingdom.
}
\note{
To download new data, log into the website, select all items of all
levels and download the data 'column-orientated csv'.

Variables are labels with the item code but full names can be found in
the source data
}
\author{
Pascal Führlich, Robin Hasse
}
