% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDisaggregate.R
\name{toolDisaggregate}
\alias{toolDisaggregate}
\title{Disaggregate energy demand by carriers into end uses by each carrier}
\usage{
toolDisaggregate(
  data,
  enduseShares,
  exclude = NULL,
  dataDisagg = NULL,
  regionMapping = NULL,
  lowerBounds = NULL
)
}
\arguments{
\item{data}{data.frame with energy demand data differentiated by carriers
that is to be disaggregated}

\item{enduseShares}{data.frame with end use shares that have to be met. These
shares might be given at aggregated regional resolution if
\code{regionmapping} is provided.}

\item{exclude}{data.frame with the columns \code{carrier} and
\code{enduse} that should list all  combinations of the two that are
excluded. All remaining combinations of the carriers in \code{data} and the
end uses in \code{enduseShares} are considered.}

\item{dataDisagg}{data.frame similar to \code{data} but already disaggregated
by carriers and end uses. The average distribution of its disaggregation
will be used as the target distribution for the minisation.}

\item{regionmapping}{data.frame with the columns \code{region} and
\code{regionAgg} that maps the regions between \code{data} and
\code{enduseShares}.}

\item{lowerBounds}{data.frame with existing energy demand data differentiated
by carriers and enduses that serve as lower bounds for the disaggregation}
}
\description{
This function takes energy demand data that differentiates only by carrier
and disaggregates it further such that all relevant combinations of carrier
and end use are differentiated. It ensures that given end use shares (across
all carriers) are met.
}
\details{
The function essentially performs a quadratic optimisation. The constraints
make sure that both the carrier quantities and the end use shares are met.
This generally leaves infinite solutions (in problematic cases none). To get
to one disaggregation, the deviation from a given distribution across all
relevant combinations of carriers and end uses is minimised.
}
\author{
Hagen Tockhorn, Robin Hasse
}
