% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDisaggregate.R
\name{toolDisaggregate}
\alias{toolDisaggregate}
\title{Disaggregates Enduse and Carrier Shares into combined enduse-carrier shares
via a quadratic optimization for an under-determined linear set of equations.}
\usage{
toolDisaggregate(data, sharesEU, sharesEC, showCheck = TRUE)
}
\arguments{
\item{data}{data.frame containing enduse-carrier shares for first estimation
in optimization problem. Columns 'enduse', 'carrier' and 'value' are required.}

\item{sharesEU}{data.frame containing aggregated enduse shares}

\item{sharesEC}{data.frame containing aggregated carrier shares}
}
\description{
Disaggregates Enduse and Carrier Shares into combined enduse-carrier shares
via a quadratic optimization for an under-determined linear set of equations.
}
\author{
Hagen Tockhorn
}
