% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolUnitConversion.R
\name{toolUnitConversion}
\alias{toolUnitConversion}
\title{convert unit}
\usage{
toolUnitConversion(x, unitConversion, dim = 3.1, removeUnit = FALSE)
}
\arguments{
\item{x}{magpie object}

\item{unitConversion}{data.frame with the columns from, to, factor}

\item{dim}{variable dimension that contains the unit as suffix}

\item{removeUnit}{boolean remove unit after conversion for clean variable
names}
}
\description{
Takes MAgPIE object with a variable dimension that has the unit as suffix
separated with '_' and converts the unit according to a unit conversion
table.
}
\author{
Robin Hasse
}
